refactor(validate): move stack-specific gate diagnostic out of neutral core
Iteration 2: resolve stack_neutrality convention violations. The pytest-specific
diagnostic pass (default command template + output parsing) now lives in a new
theforge.gate_diagnostics module, outside the stack-neutral core (coordinator/,
task/, sprint/, and the scanned config types). The coordinator calls
run_gate_diagnostic_pass() and references only configured values; the RCA packet
text no longer hardcodes tool names. Config exposes gate_diagnostic_command as an
operator override (default template resolved in the diagnostics module) instead
of a pytest-named base field.
