+++ src/theforge/coordinator/gate.py
@@ -2,16 +2,14 @@
 
 from __future__ import annotations
 
-import re
 import shlex
 from pathlib import Path
 from unittest.mock import Mock
 
 from theforge.config import ForgeConfig
-from theforge.coordinator.state import GateDebugTelemetry, GateDiagnosticTelemetry
+from theforge.coordinator.state import GateDebugTelemetry
 from theforge.task import TaskStory
 from theforge.traces import write_trace
-from theforge.workspace_env import build_workspace_env
 
 from . import util as _cu
 
@@ -173,114 +171,6 @@ def _run_gate_debug_command(
         ran=True,
         timeout_s=debug_timeout,
         exit_code=exit_code,
-        output_tail=output_tail,
-        output_truncated=len(output) > len(output_tail),
-    )
-
-
-# pytest-timeout emits the offending node id in its per-test timeout banner and
-# again in the terminal summary. Match both shapes; the summary line is the most
-# reliable, so it is scanned first.
-_PYTEST_TIMEOUT_SUMMARY_RE = re.compile(
-    r"^(?:FAILED|ERROR)\s+(?P<node>\S+::\S+)\b.*?Timeout", re.MULTILINE
-)
-_PYTEST_TIMEOUT_BANNER_RE = re.compile(r"(?P<node>\S+::\S+)\s*\n[+~]+\s*Timeout", re.MULTILINE)
-
-
-def _extract_hanging_test(output: str) -> str | None:
-    """Best-effort extraction of the single test node that exceeded the per-test timeout.
-
-    Returns the node id (e.g. ``tests/test_x.py::test_hang``) when exactly one
-    test is implicated by pytest-timeout's output, else None. When more than one
-    distinct node timed out, returns None: the "single hanging test" highlight in
-    the dev retry packet is only meaningful when the diagnostic pass isolated one.
-    """
-    nodes: list[str] = []
-    for match in _PYTEST_TIMEOUT_SUMMARY_RE.finditer(output):
-        nodes.append(match.group("node"))
-    if not nodes:
-        for match in _PYTEST_TIMEOUT_BANNER_RE.finditer(output):
-            nodes.append(match.group("node"))
-    distinct = list(dict.fromkeys(nodes))
-    if len(distinct) == 1:
-        return distinct[0]
-    return None
-
-
-def _resolve_diagnostic_test_target(config: ForgeConfig, task: TaskStory | None) -> str:
-    """Resolve the pytest target for the diagnostic pass (task override or default)."""
-    default_target = config.validation.default_test_target or "."
-    return (task.test_target if task is not None else None) or default_target
-
-
-def _run_gate_diagnostic_pass(
-    config: ForgeConfig,
-    workspace_path: Path,
-    *,
-    task: TaskStory | None,
-    iter_num: int,
-) -> GateDiagnosticTelemetry | None:
-    """Run a serialized pytest diagnostic pass after a gate timeout (issue #1217).
-
-    Runs ``<pytest_base> -n 0 --timeout=<per_test> --timeout-method=thread <target>``
-    with a hard wall-clock budget so it cannot itself become a wall-clock failure.
-    The original gate process group is already killed by ``_run_shell_detailed``
-    on timeout before this runs. faulthandler is enabled via the environment so
-    C-level frames appear alongside pytest-timeout's Python stack dump.
-
-    Returns telemetry (written to the audit + a trace file) or None when the
-    diagnostic pass is disabled.
-    """
-    val = config.validation
-    if not val.gate_diagnostic_enabled:
-        return None
-
-    per_test_timeout = val.gate_diagnostic_per_test_timeout
-    budget = val.gate_diagnostic_budget
-    test_target = _resolve_diagnostic_test_target(config, task)
-    diagnostic_cmd = (
-        f"{val.gate_diagnostic_pytest_base} -n 0 --timeout={per_test_timeout}"
-        f" --timeout-method=thread {test_target}"
-    )
-    _cu._log(f"  Running gate diagnostic pass after timeout: {diagnostic_cmd}")
-
-    env = build_workspace_env(workspace_path, extra={"PYTHONFAULTHANDLER": "1"})
-    run_shell = _cu._run_shell_detailed
-    if isinstance(run_shell, Mock):
-        ok, output, exit_code, timed_out = run_shell(
-            diagnostic_cmd,
-            workspace_path,
-            timeout=budget,
-        )
-    else:
-        ok, output, exit_code, timed_out = run_shell(
-            diagnostic_cmd,
-            workspace_path,
-            timeout=budget,
-            env=env,
-        )
-
-    if timed_out:
-        _cu._log(f"  Gate diagnostic pass hit its {budget}s budget before finishing")
-    hanging_test = _extract_hanging_test(output)
-    if hanging_test:
-        _cu._log(f"  Gate diagnostic pass isolated hanging test: {hanging_test}")
-
-    tail_chars = val.gate_output_tail_chars
-    output_tail = output[-tail_chars:]
-    write_trace(
-        workspace_path / ".forge/traces" / f"{iter_num}-gate-diagnostic.txt",
-        output,
-    )
-    return GateDiagnosticTelemetry(
-        iteration=iter_num,
-        command=diagnostic_cmd,
-        ran=True,
-        budget_s=budget,
-        per_test_timeout_s=per_test_timeout,
-        exit_code=exit_code,
-        timed_out=timed_out,
-        hanging_test=hanging_test,
         output_tail=output_tail,
         output_truncated=len(output) > len(output_tail),
     )
